LayoutTests/http/tests/performance/performance-resource-timing-cached-entries-expected.txt

CONSOLE MESSAGE: line 9: current entries: 0
PASS foundResources.length is 2
